+++ src/App.vue
@@ -1,60 +1,24 @@
 <script lang="tsx">
-import { Vue } from 'vue-class-component'
-import DefaultTable from './ui/default-table/default-table'
-import MultipleRelationSelect from './ui/relation-select/multiple-relation-select'
-import AutomaticErrorPopup from './ui/automatic-error-popup'
-import axios from './api'
+import { defineComponent, ref, Ref } from 'vue'
+import CalendarInput from './ui/calendars/calendar-input'
+import { Moment } from 'moment'
 
-export default class App extends Vue {
-  text: string = ''
-  item: any = { id: '', name: '' }
-  showPopup: boolean = false
-  popupMessage: string = ''
+export default defineComponent({
+  setup () {
+    const day: Ref<Moment|null> = ref(null)
 
-  headers = [
-    { name: 'Имя' }
-  ]
-
-  render () {
-    const line = (item: any) => {
+    return () => {
       return (
-        <div class="apptimizm-ui-default-table-row">
-          <div class="apptimizm-ui-default-table-cell">
-            {item.name}
-          </div>
-        </div>
+        <CalendarInput
+          modelValue={day.value}
+          onValueChange={(v) => { day.value = v }}
+          placeholder='Дата'
+          clearable={true}
+        />
       )
     }
-    return (
-      <div>
-        <MultipleRelationSelect
-          axios={axios}
-          endpoint='https://api.instantwebtools.net/v11/passenger'
-          itemConverter={(i: any) => ({ id: i._id, name: i.name })}
-          preselected={true}
-          modelValue={this.item}
-          onValueChange={(i: any) => { this.item = i }}
-          responseItemsKey='data'
-          responseTotalKey='totalPages'
-          requestPageKey='page'
-          requestPerPageKey='size'
-        />
-        <DefaultTable
-          axios={axios}
-          endpoint='https://api.instantwebtools.net/v1/passenger'
-          itemConverter={(i: any) => ({ id: i._id, name: i.name })}
-          headers={this.headers}
-          line={line}
-          responseItemsKey='data'
-          responseTotalKey='totalPages'
-          requestPageKey='page'
-          requestPerPageKey='size'
-        />
-        <AutomaticErrorPopup/>
-      </div>
-    )
   }
-}
+})
 </script>
 
 <style lang="sass">